Enter and Exit the Isolate whenever a lock is acquired
Whenever a lock is acquired, the Isolate is entered, and when the
lock is released, the Isolated is exited. Enter / Exit appear
to setup thread specific settings, so making sure we are
Entered, may help with some multi-threading issues we are seeing.

It is important that we only have one Locker per Isolate (something
we had before, since the V8Locker didn't have a public constructor).
Also, it's important that the lock is acquired before release
is called (also something that should have been true before, since
release checks this).

Change-Id: I3a2497b4e6c63caeb5209d5b69c6e0057c207b3d